# Improve performance of _load_custom_shape function