@jakobj 2016-12-11
my bad, I though the code said `counter += 1`. nice solution.